+++ packages/bruno-app/src/components/RequestPane/Script/index.js
@@ -6,6 +6,7 @@ import { updateRequestScript, updateResponseScript } from 'providers/ReduxStore/
 import { sendRequest, saveRequest } from 'providers/ReduxStore/slices/collections/actions';
 import { useTheme } from 'providers/Theme';
 import StyledWrapper from './StyledWrapper';
+import { updateCodeMirrorsHeight } from 'utils/common/codemirror';
 
 const Script = ({ item, collection }) => {
   const dispatch = useDispatch();
@@ -14,31 +15,9 @@ const Script = ({ item, collection }) => {
 
   const { displayedTheme } = useTheme();
   const preferences = useSelector((state) => state.app.preferences);
-  const isResponsePaneDockedToBottom = useSelector(
-    (state) => state.app.preferences.userInterface.responsePaneDockedToBottom
-  );
-
-  const updateCodeMirrorHeight = (parentId, offsetTop, dockRightHeight) => {
-    const codeMirror = document.querySelectorAll(parentId + ' .CodeMirror');
-    const pane = document.querySelector('.request-pane');
-    if (codeMirror !== null && pane !== null) {
-      codeMirror.forEach((control) => {
-        let newHeight;
-
-        if (isResponsePaneDockedToBottom) {
-          newHeight = (pane.offsetHeight - offsetTop) / 2 + 'px';
-        } else {
-          newHeight = dockRightHeight;
-        }
-        if (newHeight !== control.style.height) {
-          control.style.height = newHeight;
-        }
-      });
-    }
-  };
 
   useEffect(() => {
-    updateCodeMirrorHeight('#request-script-tab', 125, 'calc((100vh - 280px) / 2');
+    updateCodeMirrorsHeight('#request-script-tab', 125, 'calc((100vh - 280px) / 2');
   });
 
   const onRequestScriptEdit = (value) => {
